Update energy-eval endpoint to take in filename and script
The filename points to the APK file to evaluate. This should allow
it to handle multiple concurrnt requests as well. multer has been
changed so that the extension of the file is appended to the end of
the filename too. The call to Orka now passes the filename and when
Orka has finished it will delete the apk file. Will need to add
the script to the parameter as well when the frontend has implemented
ability to upload monkeyrunner scripts.