Add support for inserting energy results to backend
Categories now passed to CSVData function which calls a db function
to insert the energy results into the PostgresSQL backend

The db connects using an environmental variable DATABASE_URL. This can
be placed inside a .env file.

Categories.js file is used to implement the mapping between a category
and its index value.