Refactor energy evaluation in server.js to energyEvaluator.js
Major refactoring of server.js to store the energy evaluation code
inside a new file energyEvaluator.js. This will make it easier to
maintain the codebase and also some duplication has been removed as
well. This is in preparation for changing the get request to a new post
request now. Also:
- Nodejs has a default timeout of around 2-3 minutes for a request
  before it will attempt to perform the request again. This causes
  multiple Orka processes to try to run when using Droidmate, as
  Droidmate typically takes around 3 minutes to complete. A solution so
  far has been to increase the timeout of the server located in the main
  server.js file.