Change batch requests to perform requests in parallel
batchRequests.js updated so that requests are performed in parallel. This is
achieved by opening multiple AVDs and using a queue to store which apps are being
evaluated. The parallelism is achieved by firing each asynchronous tasks
first and collecting the promises into a list. We iterate and await each promise
in order to ensure the results are collected in the correct order.

function getAvailableAVDs() will run an android command to retrieve the list of available
AVDs that are installed on the machine. For each AVD, we assign it a port
number starting from 5554. This then is incremented by 2 each time as
emulator consoles should be even numbered and adb connections odd
numbered. The list is then stored in a global variable availableAVDs.

For each request, it takes the first element from the AVD list and 
will use that for its simulation.
Accesses to this list is mutually exclusive which is accomplished by using
locks.

Each app is also appended to a requests queue. This is needed so that we
can keep track of which apps that are being tested. This is because if two
apps have the same name, then Orka will try to override the files in
parallel. Thus, to prevent this we use a mutually exclusive queue to store
which requests are being evaluated at the moment. Accesses to the list
also uses locks to prevent concurrency issues. The request will only be
appended onto the list when the app name does not exist in the list.
When an energy evaluation has finished, the app will be removed from
the list. For requests that share the same app name this results in the apps
being evaluated sequentially. async-retry package is used to retry the append 
action at set intervals. This ensures that the the Javascript does not get blocked
by while loops or blocking events, which is needed as Javascript is
single-threaded.

In order to tell Orka which AVD and port to use, these are passed in as command
line parameters.